AccountManagerにUserExtrasの操作があるのはおかしいので分離
