+++ Yukari/src/main/java/shibafu/yukari/database/AccountManagerImpl.java
@@ -44,7 +44,6 @@ public class AccountManagerImpl implements AccountManager {
     private final DefaultVisibilityCache defaultVisibilityCache;
 
     private final List<AuthUserRecord> users = new ArrayList<>();
-    private final List<UserExtras> userExtras;
 
     public AccountManagerImpl(Context context, CentralDatabase database, ApiCollectionProvider apiCollectionProvider, StreamCollectionProvider streamCollectionProvider, DefaultVisibilityCache defaultVisibilityCache) {
         this.context = context;
@@ -54,7 +53,6 @@ public class AccountManagerImpl implements AccountManager {
         this.defaultVisibilityCache = defaultVisibilityCache;
 
         reloadUsers(true);
-        userExtras = database.getRecords(UserExtras.class, new Class[]{Collection.class}, users);
     }
 
     @Override
@@ -268,58 +266,6 @@ public class AccountManagerImpl implements AccountManager {
         }
 
         return found;
-    }
-
-    @Override
-    public void setColor(String url, int color) {
-        UserExtras extras = null;
-        for (UserExtras userExtra : userExtras) {
-            if (userExtra.getId().equals(url)) {
-                userExtra.setColor(color);
-                extras = userExtra;
-                break;
-            }
-        }
-        if (extras == null) {
-            extras = new UserExtras(url);
-            extras.setColor(color);
-            userExtras.add(extras);
-        }
-        database.updateRecord(extras);
-    }
-
-    @Override
-    public void setPriority(String url, AuthUserRecord userRecord) {
-        UserExtras extras = null;
-        for (UserExtras userExtra : userExtras) {
-            if (userExtra.getId().equals(url)) {
-                userExtra.setPriorityAccount(userRecord);
-                extras = userExtra;
-                break;
-            }
-        }
-        if (extras == null) {
-            extras = new UserExtras(url);
-            extras.setPriorityAccount(userRecord);
-            userExtras.add(extras);
-        }
-        database.updateRecord(extras);
-    }
-
-    @Override
-    @Nullable
-    public AuthUserRecord getPriority(String url) {
-        for (UserExtras userExtra : userExtras) {
-            if (userExtra.getId().equals(url)) {
-                return userExtra.getPriorityAccount();
-            }
-        }
-        return null;
-    }
-
-    @Override
-    public List<UserExtras> getUserExtras() {
-        return userExtras;
     }
 
     private static class FetchDefaultVisibilityTask extends SimpleAsyncTask {
